requirements/base.txt
Django==4.2.9
django-environ==0.11.2
python-decouple==3.8

psycopg2-binary==2.9.9

djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
django-filter==23.5
django-cors-headers==4.3.1

django-redis==5.4.0
redis==5.0.1

celery==5.3.4
django-celery-beat==2.7.0

google-generativeai==0.3.2
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
SQLAlchemy==2.0.25
sentence-transformers==2.2.2
chromadb==0.4.22

httpx==0.26.0

PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
openpyxl==3.1.2
python-pptx==0.6.23
Pillow==10.1.0
pydub==0.25.1
SpeechRecognition==3.10.1

python-dateutil==2.8.2
